refactor(workstation): extract 4 medium surfaces from inkRuntime (#890 phase 5a.2)
Second batch of surface extractions, building on #897 (5a.1). Pulls 4
more surfaces out of inkRuntime.ts into per-surface modules under
src/workstation/surfaces/. Each remained near-verbatim — the only call-
site changes are the inline `truncate` alias becoming `truncateCells`
and the URL helpers moving with the tags surface (see below).

Surfaces moved (4):
  - compose       → src/workstation/surfaces/compose/index.ts        (130)
  - branches      → src/workstation/surfaces/branches/index.ts       (133)
  - tags          → src/workstation/surfaces/tags/index.ts           (114)
  - pullRequest   → src/workstation/surfaces/pullRequest/index.ts    (152)

Shared helpers extracted:
  - buildCommitUrl, buildRefUrl → moved into src/workstation/runtime/
    utils.ts. Both are tiny pure helpers (provider repository in →
    URL out) used by the tags surface and by other render functions
    still in inkRuntime (renderCommitDiffDetail, renderInspectorRefs).
    utils.ts already imports from git/, so this is a leaf-module
    expansion — no new dependency directions.

Stale debris cleaned (left over from 5a.1's deletions):
  - The orphaned reflog-surface docstring (`Promoted reflog browser
    (#781)…`) at the seam between renderTagsSurface and the new
    location of renderPullRequestSurface
  - The orphaned promoted-filter docstring (`Filter input cursor for
    the promoted views…`) below where renderPromotedFilterAffordance
    used to live

inkRuntime.ts: 5,485 → 4,999 LOC (-486 this PR; -1,040 cumulative since
phase 5a started, -17% off the original 6,039). 4 surface render
functions, 2 URL helpers, and 2 stale docstrings deleted; 14 imports
pruned that became unused after extraction (formatSortIndicator,
formatBranchDivergence, formatBranchLastTouched, formatLogInkBranches/
Compose/TagsEmpty, buildProviderUrl, all 6 PR-panel helpers,
renderPromotedFilterAffordance).

8 of 12 promoted surfaces are now in their own modules. 5a.3 covers
the harder ones — history, status, conflicts, diff, detail — which
have richer state and inspector/preview helper dependencies.

Validation: tsc clean, lint clean, 680/680 suites & 6571/6571 tests
passing (same baseline as post-5a.1). No behavior change — every
surface renders byte-identically before/after.
